pyproject.toml
[project]
name = "test-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "fastmcp>=2.11.3",
    "google-genai>=1.32.0",
    "mcp[cli]>=1.13.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "starlette>=0.47.3",
    "uvicorn>=0.35.0",
]
